% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{DBI}
\alias{DBI}
\title{DBI Backend}
\description{
DBI backend for sessions.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DBI$new()}}
\item \href{#method-set}{\code{DBI$set()}}
\item \href{#method-has}{\code{DBI$has()}}
\item \href{#method-get}{\code{DBI$get()}}
\item \href{#method-clone}{\code{DBI$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBI$new(con, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{con}}{Live DBI connection.}

\item{\code{table}}{Name of table to store sessions.
The table must exists, it must also contain the expected
columns, as well as a \code{key} column where the user token will
be stored.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Constructor
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBI$set(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Key of the \code{value}.}

\item{\code{value}}{Value to store.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Add a key-value pair to the map.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has"></a>}}
\if{latex}{\out{\hypertarget{method-has}{}}}
\subsection{Method \code{has()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBI$has(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Key of the \code{value}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Check whether a key is present in the map.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBI$get(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Key of the \code{value}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Retrieve a value given its key.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DBI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
