% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{update_statement}
\alias{update_statement}
\title{Update Statement}
\usage{
update_statement(table, key, value)
}
\arguments{
\item{table}{Name of the table.}

\item{key}{Key (token).}

\item{value}{Data.frame row to update.}
}
\description{
Create an update statement.
}
\keyword{internal}
