% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/middleware.R
\name{eburones}
\alias{eburones}
\title{Eburones}
\usage{
eburones(backend = Local$new(), callback = function(req, res) list())
}
\arguments{
\item{backend}{Storage class to keep track of callbacks.}

\item{callback}{Function to run when a callback is created or retrieved.}
}
\description{
Session middleware.
}
